Base conda environment for use on ACCRE (safe)
This environment is for use on ACCRE post conda commercialization and only relies upon the free bioconda and conda-forge repositories. It uses python v3.10.9, which is stable and is good for installing R packages, too. 

This environment runs R4.4.1, but this could be downgraded based on package needs. It does have the package, "tidyverse," installed, which is not always possible based on versioning conflicts.